captureatrip-backend/requirements/common.txt
## Common packages
django==4.1.1
djangorestframework==3.13.1
django-storages==1.13.1
python-decouple==3.6
django-cors-headers==3.13.0
django-admin-generator==2.2.1
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3
Pillow==9.2.0
django-silk==5.0.1
drf-extra-fields==3.4.0
PyPDF2==2.10.9
boto3==1.24.78